% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_contrast.R
\name{build_contrast}
\alias{build_contrast}
\title{Create all-pairwise contrast matrix}
\usage{
build_contrast(levels, ref = NULL)
}
\arguments{
\item{levels}{Character vector of treatment levels}

\item{ref}{Reference treatment level. If \code{NULL}, the last level is used.}
}
\value{
A contrast matrix of size Choose(K, 2) x K with entries -1, 0, or 1
}
\description{
Create all-pairwise contrast matrix
}
\examples{
build_contrast(c("A", "B", "C"), ref = "A")
build_contrast(c("A", "B", "C"))
}
